requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.2.0
boto3==1.26.60
botocore==1.29.60
branca==0.6.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==1.2.0
Django==3.2.16
django-auto-logout==0.5.1
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-filter==22.1
django-formtools==2.3
django-import-export==3.2.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-pandas==0.6.6
django-phonenumber-field==7.0.2
django-storages==1.13.2
et-xmlfile==1.1.0
folium==0.14.0
fonttools==4.39.4
frozenlist==1.3.3
geographiclib==2.0
geoip2==4.6.0
geopy==2.3.0
gmplot==1.4.1
gunicorn==20.1.0
html5lib==1.1
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.2
MarkupPy==1.14
MarkupSafe==2.1.2
maxminddb==2.2.0
multidict==6.0.4
numpy==1.24.1
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
pandas==2.0.1
phonenumbers==8.13.5
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
pydyf==0.6.0
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
pypdf==3.9.0
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tablib==3.4.0
tinycss2==1.2.1
typing_extensions==4.6.2
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==1.26.14
webencodings==0.5.1
whitenoise==6.3.0
xhtml2pdf==0.2.11
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.2
zopfli==0.2.2
